python/pyproject.toml
[tool.uv.workspace]
members = ["packages/*", "examples/*"]

[tool.uv.sources]
jumpstarter = { workspace = true }
jumpstarter-cli = { workspace = true }
jumpstarter-cli-admin = { workspace = true }
jumpstarter-cli-common = { workspace = true }
jumpstarter-cli-driver = { workspace = true }
jumpstarter-driver-adb = { workspace = true }
jumpstarter-driver-androidemulator = { workspace = true }
jumpstarter-driver-ble = { workspace = true }
jumpstarter-driver-can = { workspace = true }
jumpstarter-driver-composite = { workspace = true }
jumpstarter-driver-doip = { workspace = true }
jumpstarter-driver-corellium = { workspace = true }
jumpstarter-driver-dutlink = { workspace = true }
jumpstarter-driver-energenie = { workspace = true }
jumpstarter-driver-esp32 = { workspace = true }
jumpstarter-driver-flashers = { workspace = true }
jumpstarter-driver-http = { workspace = true }
jumpstarter-driver-http-power = { workspace = true }
jumpstarter-driver-gpiod = { workspace = true }
jumpstarter-driver-ridesx = { workspace = true }
jumpstarter-driver-network = { workspace = true }
jumpstarter-driver-opendal = { workspace = true }
jumpstarter-driver-power = { workspace = true }
jumpstarter-driver-pi-pico = { workspace = true }
jumpstarter-driver-probe-rs = { workspace = true }
jumpstarter-driver-pyserial = { workspace = true }
jumpstarter-driver-qemu = { workspace = true }
jumpstarter-driver-sdwire = { workspace = true }
jumpstarter-driver-tasmota = { workspace = true }
jumpstarter-driver-tmt = { workspace = true }
jumpstarter-driver-tftp = { workspace = true }
jumpstarter-driver-snmp = { workspace = true }
jumpstarter-driver-shell = { workspace = true }
jumpstarter-driver-ssh = { workspace = true }
jumpstarter-driver-uboot = { workspace = true }
jumpstarter-driver-uds = { workspace = true }
jumpstarter-driver-uds-can = { workspace = true }
jumpstarter-driver-uds-doip = { workspace = true }
jumpstarter-driver-iscsi = { workspace = true }
jumpstarter-driver-mitmproxy = { workspace = true }
jumpstarter-driver-ustreamer = { workspace = true }
jumpstarter-driver-yepkit = { workspace = true }
jumpstarter-driver-vnc = { workspace = true }
jumpstarter-driver-xcp = { workspace = true }
jumpstarter-imagehash = { workspace = true }
jumpstarter-kubernetes = { workspace = true }
jumpstarter-mcp = { workspace = true }
jumpstarter-protocol = { workspace = true }
jumpstarter-testing = { workspace = true }
hatch-pin-jumpstarter = { workspace = true }

[dependency-groups]
docs = [
    "sphinx<8.1.0",
    "myst-parser>=4.0.0",
    "sphinxcontrib-mermaid>=0.9.2",
    "furo>=2024.8.6",
    "esbonio>=0.16.4",
    "sphinx-autobuild>=2024.4.16",
    "sphinx-click>=6.0.0",
    "sphinx-substitution-extensions>=2024.10.17",
    "requests>=2.32.3",
    "sphinxcontrib-programoutput>=0.18",
    "sphinx-copybutton>=0.5.2",
    "sphinx-inline-tabs>=2023.4.21",
]
dev = [
    "ruff==0.15.8",
    "typos>=1.23.6",
    "pre-commit>=3.8.0",
    "esbonio>=0.16.5",
    "ty>=0.0.1a8",
]

[tool.ruff]
target-version = "py311"
exclude = ["packages/jumpstarter-protocol"]
line-length = 120

[tool.ruff.lint]
exclude = ["packages/jumpstarter-protocol"]
extend-select = ["I", "C", "E", "F", "W", "B"]

[tool.ruff.lint.isort]
known-local-folder = ["packages", "examples", "jumpstarter"]


[tool.typos.default]
locale = "en-us"

[tool.typos.default.extend-words]
ser = "ser"
Pn = "Pn"
mosquitto = "mosquitto"

[tool.coverage.run]
omit = ["conftest.py", "test_*.py", "*_test.py", "*_pb2.py", "*_pb2_grpc.py"]
skip_empty = true

[tool.pytest.ini_options]
addopts = "--capture=no --doctest-modules --cov --cov-report=html --cov-report=xml"

[tool.hatch.version]
source = "vcs"
